+++ src/main/java/dbdr/domain/core/messaging/service/LineMessagingService.java
@@ -1,5 +1,6 @@
 package dbdr.domain.core.messaging.service;
 
+import java.time.LocalTime;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -8,15 +9,20 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.linecorp.bot.client.LineMessagingClient;
+import com.linecorp.bot.model.PushMessage;
 import com.linecorp.bot.model.event.FollowEvent;
 import com.linecorp.bot.model.event.MessageEvent;
 import com.linecorp.bot.model.event.message.TextMessageContent;
+import com.linecorp.bot.model.message.TextMessage;
+import com.linecorp.bot.model.profile.UserProfileResponse;
 
+import dbdr.domain.careworker.service.CareworkerMessagingService;
 import dbdr.domain.careworker.service.CareworkerService;
+import dbdr.domain.guardian.service.GuardianMessagingService;
 import dbdr.domain.guardian.service.GuardianService;
 import dbdr.global.exception.ApplicationError;
 import dbdr.global.exception.ApplicationException;
-import dbdr.global.util.LineMessagingUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -28,9 +34,9 @@ public class LineMessagingService {
 	private final ObjectMapper objectMapper;
 	private final GuardianMessagingService guardianMessagingService;
 	private final CareworkerMessagingService careworkerMessagingService;
-	private final LineMessagingUtil lineMessagingUtil;
 	private final GuardianService guardianService;
 	private final CareworkerService careworkerService;
+	private final LineMessagingClient lineMessagingClient;
 
 	@Transactional
 	public void handleLineEvent(String requestBody) {
@@ -72,7 +78,15 @@ public class LineMessagingService {
 			" 최고의 요양원 서비스 돌봄다리입니다. 🤗\n" +
 			" 서비스를 시작하려면 전화번호를 다음과 같은 형식으로 입력해주시기 바랍니다. 😄\n" +
 			" 예 : 01012345678";
-		lineMessagingUtil.sendMessageToUser(userId, followMessage);
+		sendMessageToUser(userId, followMessage);
+	}
+
+	private void sendStrangerFollowMessage(String userId, String userName) {
+		String welcomeMessage =
+			" " + userName + "님, 안녕하세요! 🌸\n" +
+				" 저희 서비스는 보호자와 요양보호사를 위한 서비스입니다. \n" +
+				" 회원가입을 통해 이용해주시기 바랍니다. 😅";
+		sendMessageToUser(userId, welcomeMessage);
 	}
 
 	// 2. Message Event 처리
@@ -97,7 +111,7 @@ public class LineMessagingService {
 			String errorMessage =
 				" 입력값이 잘못되었습니다! 😅\n" +
 				" 다시 입력해주세요. 💬\n";
-			lineMessagingUtil.sendMessageToUser(userId, errorMessage);
+			sendMessageToUser(userId, errorMessage);
 		}
 	}
 
@@ -105,7 +119,7 @@ public class LineMessagingService {
 	// 사용자가 전화 번호를 입력했을 때 발생하는 이벤트 처리
 	@Transactional
 	public void handlePhoneNumberMessage(String userId, String phoneNumber) {
-		String userName = lineMessagingUtil.getUserProfile(userId).getDisplayName();
+		String userName = getUserProfile(userId).getDisplayName();
 
 		if (guardianService.findByPhone(phoneNumber) != null) {
 			guardianMessagingService.handleGuardianPhoneMessage(userId, phoneNumber);
@@ -120,34 +134,62 @@ public class LineMessagingService {
 	@Transactional
 	public void handleReservationMessage(String userId, String ampm, String hour, String minute) {
 		String confirmationMessage =
-			"감사합니다! 😊\n" +
-				"입력하신 시간 " + ampm + " " + hour + "시 " + minute + "분" + "에 알림을 보내드릴게요. 💬\n" +
-				"언제든지 알림 시간을 변경하고 싶으시면 다시 알려주세요!";
+			" 감사합니다! 😊\n" +
+				" 입력하신 시간 " + ampm + " " + hour + "시 " + minute + "분" + "에 알림을 보내드릴게요. 💬\n" +
+				" 언제든지 알림 시간을 변경하고 싶으시면 다시 알려주세요!";
 
 		if (guardianService.findByLineUserId(userId) != null) {
-			lineMessagingUtil.sendMessageToUser(userId, confirmationMessage);
+			sendMessageToUser(userId, confirmationMessage);
 			guardianMessagingService.updateGuardianAlertTime(userId, ampm, hour, minute);
 		} else if (careworkerService.findByLineUserId(userId) != null) {
-			lineMessagingUtil.sendMessageToUser(userId, confirmationMessage);
+			sendMessageToUser(userId, confirmationMessage);
 			careworkerMessagingService.updateCareworkerAlertTime(userId, ampm, hour, minute);
 		} else {
 			userFoundFailedMessage(userId);
 		}
 	}
 
-	private void sendStrangerFollowMessage(String userId, String userName) {
-		String welcomeMessage =
-			" " + userName + "님, 안녕하세요! 🌸\n" +
-				" 저희 서비스는 보호자와 요양보호사를 위한 서비스입니다. \n" +
-				" 회원가입을 통해 이용해주시기 바랍니다. 😅";
-		lineMessagingUtil.sendMessageToUser(userId, welcomeMessage);
-	}
+	// +) 기타 메서드들
 
 	private void userFoundFailedMessage(String userId) {
 		String errorMessage =
 			"등록된 정보가 확인되지 않았습니다. 😊\n" +
 				"서비스 이용을 위해 요양원에 문의하시거나, 전화번호를 다시 인증해주시기 바랍니다. ☎️\n" +
 				"예: 01012345678";
-		lineMessagingUtil.sendMessageToUser(userId, errorMessage);
+		sendMessageToUser(userId, errorMessage);
+	}
+
+	// 사용자에게 메시지를 보내는 메서드
+	public void sendMessageToUser(String userId, String message) {
+		TextMessage textMessage = new TextMessage(message);
+		PushMessage pushMessage = new PushMessage(userId, textMessage);
+
+		try {
+			lineMessagingClient.pushMessage(pushMessage).get();
+			log.info("Message sent successfully to user: {}", userId);
+		} catch (Exception e) {
+			log.error("Failed to send message to user: {}", userId, e);
+			throw new ApplicationException(ApplicationError.MESSAGE_SEND_FAILED);
+		}
+	}
+
+	// UserId를 통해 라인 사용자 프로필 정보 가져오는 메서드
+	public UserProfileResponse getUserProfile(String userId) {
+		try {
+			return lineMessagingClient.getProfile(userId).get();
+		} catch (Exception e) {
+			log.error("Failed to get user profile: {}", userId, e);
+			throw new ApplicationException(ApplicationError.FAILED_TO_GET_USER_PROFILE);
+		}
+	}
+
+	// AM/PM 및 시간을 LocalTime으로 변환하는 메서드
+	public LocalTime convertToLocalTime(String ampm, int hour, int minute) {
+		if (ampm.equalsIgnoreCase("오후") && hour != 12) {
+			hour += 12;
+		} else if (ampm.equalsIgnoreCase("오전") && hour == 12) {
+			hour = 0;  // 오전 12시는 0시로 변환
+		}
+		return LocalTime.of(hour, minute);  // 시간과 분을 함께 설정
 	}
 }
